categoryProccess

+++ DataAccessLayer/Migrations/20240731062026_mig1.cs
@@ -44,7 +44,7 @@ namespace DataAccessLayer.Migrations
                 name: "Categories",
                 columns: table => new
                 {
-                    CategoryID = table.Column<byte>(type: "tinyint", nullable: false)
+                    CategoryID = table.Column<int>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     CategoryName = table.Column<string>(type: "nvarchar(50)", maxLength: 50, nullable: false),
                     CategoryStatus = table.Column<bool>(type: "bit", nullable: false)
